check if zshrc and bashrc exist

+++ install.py
@@ -29,16 +29,18 @@ cat_fn = input("Enter the display cat function name :")
 
 if (not os.path.isdir("{}.cats".format(home))):
 	shutil.copytree("./cats", "{}.cats".format(home))
-shutil.copy("{}.zshrc".format(home), "{}.zshrc_backup".format(home))
-shutil.copy("{}.bashrc".format(home), "{}.bashrc_backup".format(home))
 
-with open("{}.bashrc".format(home), 'a') as file:
-	file.write('\n\n# Addition by shell-cat\n')
-	file.write(PUSH_FN.replace("$$$1", push_fn))
-	file.write("\n")
-	file.write(CAT_FN.replace("$$$1", cat_fn))
+if (os.path.exists("{}.bashrc".format(home))):
+	shutil.copy("{}.bashrc".format(home), "{}.bashrc_backup".format(home))
+	with open("{}.bashrc".format(home), 'a') as file:
+		file.write('\n\n# Addition by shell-cat\n')
+		file.write(PUSH_FN.replace("$$$1", push_fn))
+		file.write("\n")
+		file.write(CAT_FN.replace("$$$1", cat_fn))
 
-with open("{}.zshrc".format(home), 'a') as file:
+if (os.path.exists("{}.zshrc".format(home))):
+	shutil.copy("{}.zshrc".format(home), "{}.zshrc_backup".format(home))
+	with open("{}.zshrc".format(home), 'a') as file:
 	file.write('\n\n# Addition by shell-cat\n')
 	file.write(PUSH_FN.replace("$$$1", push_fn))
 	file.write("\n")
